chore: remove unnecessary conversions

+++ block/manager.go
@@ -337,7 +337,7 @@ func (m *Manager) SyncLoop(ctx context.Context, cancel context.CancelFunc) {
 			block := blockEvent.Block
 			daHeight := blockEvent.DAHeight
 			blockHash := block.Hash().String()
-			blockHeight := uint64(block.Height())
+			blockHeight := block.Height()
 			m.logger.Debug("block body retrieved",
 				"height", blockHeight,
 				"daHeight", daHeight,
@@ -398,7 +398,7 @@ func (m *Manager) trySyncNextBlock(ctx context.Context, daHeight uint64) error {
 			return nil
 		}
 
-		bHeight := uint64(b.Height())
+		bHeight := b.Height()
 		m.logger.Info("Syncing block", "height", bHeight)
 		// Validate the received block before applying
 		if err := m.executor.Validate(m.lastState, b); err != nil {
@@ -417,7 +417,7 @@ func (m *Manager) trySyncNextBlock(ctx context.Context, daHeight uint64) error {
 			return fmt.Errorf("failed to Commit: %w", err)
 		}
 
-		err = m.store.SaveBlockResponses(uint64(bHeight), responses)
+		err = m.store.SaveBlockResponses(bHeight, responses)
 		if err != nil {
 			return fmt.Errorf("failed to save block responses: %w", err)
 		}
